Changed image transfer protocol

+++ app/src/main/java/edu/mtu/cs3421/voto/UDPclient.java
@@ -236,7 +236,7 @@ public class UDPclient {
     class MediaResponse{
         int imgLength;
         byte imgID;
-        byte packetCount;
+        int packetCount;
         MediaResponse(){
 
         }
